Simplify code and remove redundant docstrings
Removed extensive docstrings from functions in `udf_catboost_inference.py` to streamline the implementation and improve readability. Also pruned code and markdown sections from the associated Jupyter notebook to focus on the key processing tasks.
